Moved shouldShowWebinarPromoNotification to an helper
